Fix: Dataset: When configuring the "general chunk method," options such as chunk size and parent-child slicing are unavailable.
