+++ web/src/hooks/use-user-setting-request.tsx
@@ -2,7 +2,7 @@ import message from '@/components/ui/message';
 import { Modal } from '@/components/ui/modal/modal';
 import { ResponseGetType } from '@/interfaces/database/base';
 import { IToken } from '@/interfaces/database/chat';
-import { ITenantInfo } from '@/interfaces/database/knowledge';
+import { ITenantInfo } from '@/interfaces/database/dataset';
 import { ILangfuseConfig } from '@/interfaces/database/system';
 import {
   ITenant,
@@ -213,6 +213,7 @@ export const useFetchSystemVersion = () => {
         setLoading(false);
       }
     } catch (error) {
+      console.warn(error);
       setLoading(false);
     }
   }, []);
