+++ web/src/interfaces/database/agent.ts
@@ -33,7 +33,7 @@ export interface ISwitchForm {
 import { AgentCategory } from '@/constants/agent';
 import { Edge, Node } from '@xyflow/react';
 import { IReference, Message } from './chat';
-import { IKnowledge } from './knowledge';
+import { IDataset } from './dataset';
 
 export type DSLComponents = Record<string, IOperator>;
 
@@ -81,7 +81,7 @@ export declare interface IFlow {
   release?: boolean;
   release_time?: number;
   last_publish_time?: number;
-  datasets?: Pick<IKnowledge, 'id' | 'name' | 'avatar'>[];
+  datasets?: Pick<IDataset, 'id' | 'name' | 'avatar'>[];
 }
 
 export interface IFlowTemplate {
