+++ web/src/pages/next-search/document-preview-modal/index.tsx
@@ -7,7 +7,7 @@ import {
 } from '@/hooks/use-document-request';
 import { IModalProps } from '@/interfaces/common';
 import { IReferenceChunk } from '@/interfaces/database/chat';
-import { IChunk } from '@/interfaces/database/knowledge';
+import { IChunk } from '@/interfaces/database/dataset';
 import { cn } from '@/lib/utils';
 import { useEffect, useState } from 'react';
