+++ web/src/utils/document-util.ts
@@ -1,7 +1,7 @@
 import { Images, SupportedPreviewDocumentTypes } from '@/constants/common';
 import { UploadFile } from '@/interfaces/antd-compat';
 import { IReferenceChunk } from '@/interfaces/database/chat';
-import { IChunk } from '@/interfaces/database/knowledge';
+import { IChunk } from '@/interfaces/database/dataset';
 import { get } from 'lodash';
 import { v4 as uuid } from 'uuid';
